@@ -1,6 +1,7 @@
 {
   "name": "recipe-website",
   "version": "0.1.0",
+  "type": "module",
   "private": true,
   "prisma": {
     "seed": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seed.ts"
@@ -20,44 +21,47 @@
     "@aws-sdk/s3-request-presigner": "^3.515.0",
     "@dnd-kit/core": "^6.0.8",
     "@dnd-kit/sortable": "^7.0.2",
-    "@formkit/auto-animate": "^1.0.0-beta.3",
+    "@formkit/auto-animate": "^0.8.1",
     "@headlessui/react": "^1.7.3",
     "@heroicons/react": "^2.0.11",
     "@hookform/error-message": "^2.0.1",
-    "@hookform/resolvers": "^2.9.8",
+    "@hookform/resolvers": "^3.3.4",
     "@next-auth/prisma-adapter": "^1.0.5",
-    "@prisma/client": "^5.0.0",
+    "@prisma/client": "^5.6.0",
+    "@t3-oss/env-nextjs": "^0.9.2",
     "@tanstack/react-query": "^4.29.23",
     "@tanstack/react-query-devtools": "^4.29.23",
-    "@trpc/client": "^10.34.0",
-    "@trpc/next": "^10.34.0",
-    "@trpc/react-query": "^10.34.0",
-    "@trpc/server": "^10.34.0",
+    "@trpc/client": "^10.45.1",
+    "@trpc/next": "^10.45.1",
+    "@trpc/react-query": "^10.45.1",
+    "@trpc/server": "^10.45.1",
     "class-variance-authority": "^0.6.1",
     "framer-motion": "^11.0.5",
-    "next": "^13.1.6",
-    "next-auth": "^4.22.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.1.2",
+    "next-auth": "^4.24.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.36.0",
     "react-icons": "^5.0.1",
+    "server-only": "^0.0.1",
     "superjson": "^1.12.4",
     "uuid": "^9.0.1",
-    "zod": "^3.21.4"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
     "@faker-js/faker": "^7.3.0",
-    "@next/eslint-plugin-next": "^13.0.0",
-    "@types/node": "^18.0.0",
-    "@types/react": "18.2.62",
-    "@types/react-dom": "18.2.15",
+    "@types/eslint": "^8.56.2",
+    "@types/node": "^20.11.25",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.20",
     "@types/uuid": "^9.0.8",
-    "autoprefixer": "^10.4.16",
-    "eslint": "8.56.0",
-    "eslint-config-next": "14.1.0",
+    "@typescript-eslint/eslint-plugin": "^7.0.0",
+    "@typescript-eslint/parser": "^7.0.0",
+    "eslint": "^8.56.0",
+    "eslint-config-next": "^14.1.0",
     "postcss": "^8.4.14",
-    "prettier": "2.7.1",
-    "prettier-plugin-tailwindcss": "^0.1.13",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.12",
     "prisma": "^5.6.0",
     "tailwindcss": "^3.3.3",
     "ts-node": "^10.9.2",
